Refactors style files in 'global' to respective levels and components in 'app'
