+++ src/app/about/profile/Profile.jsx
@@ -1,9 +1,10 @@
 import React from 'react'
-import { IconMed } from '../../../global/components/IconMed'
+import { IconMed } from '../../../global/components/icon/IconMed'
 import headShotImage from './img/head-shot'
 import developerIcon from './img/developer'
 import projectManagerIcon from './img/project-manager'
 import analystIcon from './img/analyst'
+import './style/profile'
 
 export const Profile = () => {
   return (
