+++ src/app/portfolio/Portfolio.jsx
@@ -13,6 +13,7 @@ import { Applications } from './applications/Applications'
 import featuredContentData from '../../global/data/featured-content'
 import blogPostsData from '../../global/data/blogPosts/blog-posts-list'
 import applicationsData from '../../global/data/applications'
+import './style/portfolio'
 
 const routerParams = {
   featured: {
