+++ src/app/portfolio/blog/BlogPost.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useParams } from 'react-router-dom'
-import { NavLinkUpRouteTop } from '../../../global/components/NavLinkUpRouteTop'
-import { WebLink } from '../../../global/components/WebLink'
+import { NavLinkUpRouteTop } from '../../../global/components/link/NavLinkUpRouteTop'
+import { WebLink } from '../../../global/components/link/WebLink'
 import './style/blog-post'
 
 export const BlogPost = ({ blogPostsList, blogPostsDataPath }) => {
