+++ src/app/portfolio/featured/Featured.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { FeaturedContentCard } from './FeaturedContentCard'
 import featuredContentData from '../../../global/data/featured-content'
+import './style/featured'
 
 export const Featured = () => {
   const featuredContentItems = featuredContentData.map(featuredContentitem => {
