+++ src/app/top/App.jsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import { Route, Redirect, Switch } from 'react-router-dom'
 import { Footer } from './Footer'
-import { NavBar } from './navbar/Navbar'
-import { About } from './about/About'
-import { Portfolio } from './portfolio/Portfolio'
-import { EmailForm } from './contact/EmailForm'
+import { NavBar } from './Navbar'
+import { About } from '../about/About'
+import { Portfolio } from '../portfolio/Portfolio'
+import { EmailForm } from '../contact/EmailForm'
+import './style/global'
 
 const routesParams = {
   about: {
