--- src/global/css/common.scss
@@ -1,7 +1,0 @@
-.icon-small {
-  padding: 5px;
-}
-
-.icon-small-plus {
-  padding: 0 1px 0 1px;
-}
