+++ src/index.js
@@ -1,13 +1,9 @@
 import React from "react"
 import ReactDOM from "react-dom"
 import { BrowserRouter } from "react-router-dom"
-import { App } from './app/App'
-import "./global/css/common"
-import "./global/css/global"
-import "./global/css/portfolio"
-import "./global/css/project-card"
-import "./global/css/profile"
-import "./global/css/social-media"
+import { App } from './app/top/App'
+import "./global/components/icon/style/icon"
+import "./global/components/link/style/link"
 import './global/fonts/arial'
 
 ReactDOM.render(<BrowserRouter> <App /> </BrowserRouter>,
